+++ frontend/src/app/(posts)/posts/edit/page.tsx
@@ -3,7 +3,7 @@ import { api } from '@/util/api';
 import { PostType } from '@/type/community/type';
 import dynamic from 'next/dynamic';
 
-export const Writer = dynamic(() => import('./writer'), {
+const Writer = dynamic(() => import('@/app/(common)/(aboutContent)/(content)/writer'), {
   ssr: false,
 });
 
@@ -23,10 +23,8 @@ export default async function Page(page: SearchParams) {
   return (
     <section className="flex flex-col gap-2">
       <h2 className="text-lightBlue font-bold border-b-4 border-blue">게시글 수정</h2>
-      <div className="flex justify-between">
-        <input value={data.title} disabled className="text-black/30" />
-      </div>
-      <Writer {...data} />
+      <input value={data.title} disabled className="text-black/30" />
+      <Writer data={data} />
     </section>
   );
 }
